Struct frounding::RoundingState

pub struct RoundingState {
    // some fields omitted
}

Keep track of the current CPU rounding mode (both the SSE and the x87 FPU units).

Methods

impl RoundingState

fn new() -> Self

Create new RoundingState struct and fill it with the current status of control regiters. When going out of scope the initial rounding mode is set.

fn current_rounding(&self) -> (RoundingMode, RoundingMode)

Return the current rounding mode of both SSE and FPU.

fn sse_rounding(&self) -> RoundingMode

Return the current rounding mode of the SSE unit.

fn fpu_rounding(&self) -> RoundingMode

Return the current rounding mode of x87 FPU unit.

fn upward(&mut self)

Set the rounding mode toward +infinity for both SSE and x87 FPU.

fn sse_upward(&mut self)

Set the rounding mode of the SSE toward +infinity.

fn fpu_upward(&mut self)

Set the rounding mode of the FPU toward +infinity.

fn downward(&mut self)

Set the rounding mode toward -infinity for both SSE and x87 FPU.

fn sse_downward(&mut self)

Set the rounding mode of the SSE toward -infinity.

fn fpu_downward(&mut self)

Set the rounding mode of the FPU toward -infinity.

fn to_nearest(&mut self)

Set the rounding mode to nearest for both SSE and x87 FPU.

fn sse_to_nearest(&mut self)

Set the rounding mode of the SSE to nearest

fn fpu_to_nearest(&mut self)

Set the rounding mode of FPU to nearest

fn to_zero(&mut self)

Set the rounding mode toward zero (aka truncate) for both SSE and x87 FPU.

fn sse_to_zero(&mut self)

Set the rounding mode of SSE toward zero (aka truncate)

fn fpu_to_zero(&mut self)

Set the rounding mode of FPU toward zero (aka truncate)

fn reset(&mut self)

Reset the rounding mode to the default one (the one found when calling the new function) for both FPU and SEE.

fn sse_reset(&mut self)

Reset the rounding mode to the default one (the one found when calling the new function) for SEE.

fn fpu_reset(&mut self)

Reset the rounding mode to the default one (the one found when calling the new function) for FPU.